+++ R/interactions.r
@@ -6,6 +6,8 @@
 #' @param y A character string of the outcome variable.
 #' @param time A character string of the time variable. If `NULL`, logistic regression is used.
 #'   Otherwise, Cox proportional hazards regression is used.
+#' @param time2 A character string of the ending time of the interval for interval censored or counting process
+#'   data only.
 #' @param predictors The predictor variables to be scanned for interactions. If `NULL`, all variables
 #'   except `y` and `time` are taken as predictors.
 #' @param group_vars The group variables to be scanned for interactions. If `NULL`, all variables
@@ -22,7 +24,7 @@
 #' @examples
 #' data(cancer, package = "survival")
 #' interaction_scan(cancer, y = "status", time = "time", save_table = FALSE)
-interaction_scan <- function(data, y, time = NULL, predictors = NULL, group_vars = NULL, covars = NULL,
+interaction_scan <- function(data, y, time = NULL, time2 = NULL, predictors = NULL, group_vars = NULL, covars = NULL,
                              try_rcs = TRUE, p_adjust_method = "BH", save_table = FALSE, filename = NULL) {
   analysis_type <- if (!is.null(time)) {
     "cox"
@@ -32,11 +34,11 @@ interaction_scan <- function(data, y, time = NULL, predictors = NULL, group_vars
     "linear"
   }
   if (is.null(predictors)) {
-    predictors <- setdiff(colnames(data), c(y, time))
+    predictors <- setdiff(colnames(data), c(y, time, time2))
     message("Taking all variables as interaction predictors")
   }
   if (is.null(group_vars)) {
-    group_vars <- setdiff(colnames(data), c(y, time))
+    group_vars <- setdiff(colnames(data), c(y, time, time2))
     message("Taking all variables as group variables")
   }
   if (any(!predictors %in% colnames(data))) {
@@ -53,15 +55,18 @@ interaction_scan <- function(data, y, time = NULL, predictors = NULL, group_vars
   for (predictor in predictors) {
     for (group_var in group_vars) {
       if (group_var != predictor) {
-        nvalid <- sum(complete.cases(data[, c(time, y, predictor, group_var, covars)]))
+        nvalid <- sum(complete.cases(data[, c(time, time2, y, predictor, group_var, covars)]))
         if (nvalid < 10) {
           next
         }
 
-        p1 <- interaction_p_value(data, y, predictor, group_var, time = time, covars = covars)
+        p1 <- interaction_p_value(data, y, predictor, group_var, time = time, time2 = time2, covars = covars)
         if (try_rcs && length(unique(data[, predictor])) > 10) {
           p2 <- tryCatch(
-            interaction_p_value(data, y, predictor, group_var, time = time, covars = covars, rcs_knots = 4),
+            interaction_p_value(data, y, predictor, group_var,
+              time = time, time2 = time2, covars = covars,
+              rcs_knots = 4
+            ),
             error = function(e) {
               NA
             }
@@ -107,6 +112,8 @@ interaction_scan <- function(data, y, time = NULL, predictors = NULL, group_vars
 #'   numeric variable is provided, it will be split by the median value.
 #' @param time A character string of the time variable. If `NULL`, logistic regression is used.
 #'   Otherwise, Cox proportional hazards regression is used.
+#' @param time2 A character string of the ending time of the interval for interval censored or counting process
+#'   data only.
 #' @param covars A character vector of covariate names.
 #' @param group_colors A character vector of colors for the plot. If `NULL`, the default colors are used.
 #' @param save_plot A logical value indicating whether to save the plot.
@@ -134,9 +141,10 @@ interaction_scan <- function(data, y, time = NULL, predictors = NULL, group_vars
 #'   y = "wt.loss", predictor = "age", group_var = "sex",
 #'   save_plot = FALSE
 #' )
-interaction_plot <- function(data, y, predictor, group_var, time = NULL, covars = NULL, group_colors = NULL,
-                             save_plot = FALSE, filename = NULL, height = 4, width = 4, xlab = predictor,
-                             ylab = NULL, show_n = TRUE, group_title = group_var, ...) {
+interaction_plot <- function(data, y, predictor, group_var, time = NULL, time2 = NULL, covars = NULL,
+                             # cluster = NULL,
+                             group_colors = NULL, save_plot = FALSE, filename = NULL, height = 4, width = 4,
+                             xlab = predictor, ylab = NULL, show_n = TRUE, group_title = group_var, ...) {
   if (!is.null(time)) {
     analysis_type <- "cox"
     if (is.null(ylab)) {
@@ -159,7 +167,7 @@ interaction_plot <- function(data, y, predictor, group_var, time = NULL, covars 
   if (is.null(group_colors)) {
     group_colors <- emp_colors
   }
-  if (any(c(y, time, predictor, group_var) %in% covars)) {
+  if (any(c(y, time, time2, predictor, group_var) %in% covars)) {
     stop("Conflict of model variables!")
   }
   if (is.null(filename)) {
@@ -171,9 +179,9 @@ interaction_plot <- function(data, y, predictor, group_var, time = NULL, covars 
   }
   default_expansion <- c(0.05, 0, 0.15, 0)
 
-  dat <- dplyr::select(data, all_of(c(y, predictor, group_var, time, covars)))
-  if (".predictor" %in% c(y, time, covars)) stop("Colname '.predictor' is reserved!")
-  if (".group_var" %in% c(y, time, covars)) stop("Colname '.group_var' is reserved!")
+  dat <- dplyr::select(data, all_of(c(y, predictor, group_var, time, time2, covars)))
+  if (".predictor" %in% c(y, time, time2, covars)) stop("Colname '.predictor' is reserved!")
+  if (".group_var" %in% c(y, time, time2, covars)) stop("Colname '.group_var' is reserved!")
   colnames(dat)[c(2, 3)] <- c(".predictor", ".group_var")
   dat <- na.omit(dat)
   dat$.group_var <- to_factor(dat$.group_var)
@@ -203,11 +211,11 @@ interaction_plot <- function(data, y, predictor, group_var, time = NULL, covars 
   tryCatch(
     {
       formula <- create_formula(y, ".predictor",
-        group_var = ".group_var", time = time,
+        group_var = ".group_var", time = time, time2 = time2,
         covars = covars, interaction = TRUE
       )
       model <- fit_model(formula, data = dat, analysis_type = analysis_type, rms = TRUE)
-      p_value <- interaction_p_value(dat, y, ".predictor", ".group_var", time = time, covars = covars)
+      p_value <- interaction_p_value(dat, y, ".predictor", ".group_var", time = time, time2 = time2, covars = covars)
       y1 <- as.data.frame(Predict(model, .predictor, .group_var,
         fun = pred_fun, type = "predictions", conf.int = 0.95, digits = 2,
         ref.zero = analysis_type %in% c("cox", "logistic")
@@ -281,12 +289,12 @@ interaction_plot <- function(data, y, predictor, group_var, time = NULL, covars 
     tryCatch(
       {
         formula2 <- create_formula(y, ".predictor",
-          group_var = ".group_var", time = time, covars = covars, rcs_knots = 4,
+          group_var = ".group_var", time = time, time2 = time2, covars = covars, rcs_knots = 4,
           interaction = TRUE
         )
         model2 <- fit_model(formula2, data = dat, analysis_type = analysis_type, rms = TRUE)
         rcs_p_value <- interaction_p_value(dat, y, ".predictor", ".group_var",
-          time = time, covars = covars,
+          time = time, time2 = time2, covars = covars,
           rcs_knots = 4
         )
         y2 <- as.data.frame(Predict(model2, .predictor, .group_var,
@@ -367,6 +375,8 @@ interaction_plot <- function(data, y, predictor, group_var, time = NULL, covars 
 #'   numeric variable is provided, it will be split by the median value.
 #' @param time A character string of the time variable. If `NULL`, linear or logistic regression is used.
 #'   Otherwise, Cox proportional hazards regression is used.
+#' @param time2 A character string of the ending time of the interval for interval censored or counting process
+#'   data only.
 #' @param covars A character vector of covariate names.
 #' @param rcs_knots The number of rcs knots. If `NULL`, a linear model would be fitted instead.
 #' @returns A numerical, the interaction p-value
@@ -377,7 +387,9 @@ interaction_plot <- function(data, y, predictor, group_var, time = NULL, covars 
 #'   data = cancer, y = "status", predictor = "age", group_var = "sex",
 #'   time = "time", rcs_knots = 4
 #' )
-interaction_p_value <- function(data, y, predictor, group_var, time = NULL, covars = NULL, rcs_knots = NULL) {
+interaction_p_value <- function(
+    data, y, predictor, group_var, time = NULL, time2 = NULL, covars = NULL,
+    rcs_knots = NULL) {
   analysis_type <- if (!is.null(time)) {
     "cox"
   } else if (length(levels(as.factor(data[[y]]))) == 2) {
@@ -386,14 +398,14 @@ interaction_p_value <- function(data, y, predictor, group_var, time = NULL, cova
     "linear"
   }
   data[[group_var]] <- to_factor(data[[group_var]])
-  covars <- remove_conflict(covars, c(y, predictor, group_var, time))
+  covars <- remove_conflict(covars, c(y, predictor, group_var, time, time2))
 
   formula1 <- create_formula(y, predictor, group_var,
-    time = time, covars = covars, rcs_knots = rcs_knots,
+    time = time, time2 = time2, covars = covars, rcs_knots = rcs_knots,
     interaction = FALSE
   )
   formula2 <- create_formula(y, predictor, group_var,
-    time = time, covars = covars, rcs_knots = rcs_knots,
+    time = time, time2 = time2, covars = covars, rcs_knots = rcs_knots,
     interaction = TRUE
   )